[code] Add option to append a file to output file. (NXTHBOTIMG-98)
Added the option -a/--append-file to the COM boot image compiler.
The file is appended to the output file after after the output image has been generated and written.
It is not used during generation and is not considered for any size calculations, min. size padding or max. size checking.